requirements.txt
discord.py==1.5.1
tinydb==4.1.1